Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/browsers/the-window-object/window-open-noopener-expected.txt

CONSOLE MESSAGE: line 120: testError: special target must be one of: _self, _parent, _top

FAIL window.open() with "noopener" tests testError: special target must be one of: _self, _parent, _top
